updates to data collection methods in main_peta_dist

+++ main_peta_dist.py
@@ -1,6 +1,7 @@
 #! /usr/bin/python
 
 import math
+import os.path
 import numpy as np
 import matplotlib.pyplot as plt
 import ROOT
@@ -10,21 +11,21 @@ import Detector
 import Drawing
 
 # VIS or STATS
-mode = "VIS"
+mode = "STATS"
 visWithStats = False
 
 if mode=="VIS":
     ntrajs = 15
     trajs = []
 if mode=="STATS":
-    ntrajs = 2500
+    ntrajs = 10000
     trajs = []
 
-outname = "data/detectorHits/stats"
+outname = "data/detectorHits/stats_updown_nomsc_eta19_08101323"
 
-Params.BFieldType = 'cms'
-Params.MSCtype = 'PDG'
-Params.EnergyLossOn = True
+Params.BFieldType = 'updown'
+Params.MSCtype = 'none'
+Params.EnergyLossOn = False
 Params.Q = 1
 Params.m = 105.
 
@@ -50,41 +51,53 @@ normToDetect = normToDetect/np.linalg.norm(normToDetect)
 #center point
 center = normToDetect*distToDetect
 # y-axis (in plane)
-detVert = np.array([0,1,0])
-# another orthogonal vector to norm in plane ((normToDetect, detVert, detOrth) form an ON basis)
-detOrth = np.cross(normToDetect,detVert)
+detV = np.array([0,1,0])
+# another orthogonal vector to norm in plane ((normToDetect, detV, detW) form an ON basis)
+detW = np.cross(normToDetect,detV)
 
 detWidth = 4.
 detHeight = 4.
 
-detectorDict = {"norm":normToDetect, "dist":distToDetect, "vert":detVert, 
-                "orth":detOrth, "width":detWidth, "height":detHeight}
+detectorDict = {"norm":normToDetect, "dist":distToDetect, "v":detV, 
+                "w":detW, "width":detWidth, "height":detHeight}
 
-c1 = center + detOrth*detWidth/2 + detVert*detHeight/2
-c2 = center + detOrth*detWidth/2 - detVert*detHeight/2
-c3 = center - detOrth*detWidth/2 - detVert*detHeight/2
-c4 = center - detOrth*detWidth/2 + detVert*detHeight/2
+c1 = center + detW*detWidth/2 + detV*detHeight/2
+c2 = center + detW*detWidth/2 - detV*detHeight/2
+c3 = center - detW*detWidth/2 - detV*detHeight/2
+c4 = center - detW*detWidth/2 + detV*detHeight/2
 
 intersects = []
 
-stats = ROOT.TNtuple("stats","stats","p:pT:eta:phi:theta:thOrth:thVert")
+stats = ROOT.TNtuple("stats","stats","q:p:pT:eta:phi:theta:thW:thV:w:v")
 ntotaltrajs = 0
 
-txtfile = open(outname+".txt",'w')
-txtfile.close()
+if mode=="STATS":
+    if os.path.isfile(outname+".txt"):
+        ow = 'q'
+        while ow not in 'yYnN':
+            ow = raw_input("Overwrite file? (y/n) ")
+        if ow in 'yY':
+            txtfile = open(outname+".txt",'w')
+        else:
+            txtfile = open(outname+".txt",'a')
+    else:
+        txtfile = open(outname+".txt",'w')
+    txtfile.close()
 
 while len(trajs)<ntrajs:
     magp = ROOT.Double(1e9)
     eta = ROOT.Double(-1)
 
-    etalow = 1.2
-    etahigh = 2.5
+    etalow = 1.3
+    etahigh = 2.4
 
     while eta<etalow or eta>etahigh:
         p_eta_dist.GetRandom2(magp,eta)
 
+    eta = 1.9
+
     th = 2*np.arctan(np.exp(-eta))
-    phimin, phimax = -1.4,1.4
+    phimin, phimax = -0.6, 0.6
     phi = np.random.rand() * (phimax-phimin) + phimin
     Params.Q = np.random.randint(2)*2 - 1
 
@@ -96,7 +109,7 @@ while len(trajs)<ntrajs:
     if mode=="VIS":
         trajs.append(traj)
 
-    intersection, theta, thOrth, thVert = Detector.FindIntersection(traj, detectorDict)
+    intersection, theta, thW, thV = Detector.FindIntersection(traj, detectorDict)
     if intersection != None:
         intersects.append(intersection)
         print len(trajs), ": p =",magp, ", eta =", eta, ", phi =", phi
@@ -107,13 +120,16 @@ while len(trajs)<ntrajs:
                 trajs.append(traj)
             else:
                 trajs.append(0)
-            stats.Fill(magp,magp*np.sin(th),eta,phi,theta, thOrth, thVert)
+            w = np.dot(intersection, detW)
+            v = np.dot(intersection, detV)
+            stats.Fill(Params.Q,magp,magp*np.sin(th),eta,phi,theta, thW, thV, w, v)
             txtfile = open(outname+".txt",'a')
-            txtfile.write("{0:f}\t{1:f}\t{2:f}\t{3:f}\t{4:f}\t{5:f}\t{6:f}\n".format(magp,magp*np.sin(th),eta,phi,theta,thOrth,thVert))
+            txtfile.write("{0:d}\t{1:f}\t{2:f}\t{3:f}\t{4:f}\t{5:f}\t{6:f}\t{7:f}\t{8:f}\t{9:f}\n".format(Params.Q,magp,magp*np.sin(th),eta,phi,theta,thW,thV,w,v))
 
-fid = ROOT.TFile(outname+".root","RECREATE")
-stats.Write()
-fid.Close()
+if mode=="STATS":
+    fid = ROOT.TFile(outname+".root","RECREATE")
+    stats.Write()
+    fid.Close()
 
 print "Efficiency:", float(len(intersects))/ntotaltrajs
